+++ .gitignore
@@ -2,6 +2,7 @@
 #
 .venv
 __pycache__
+*.egg-info/
 
 #VSCode
 #
