+++ api/setup.cfg
@@ -1,0 +1,20 @@
+[metadata]
+name = mirrsearchapi
+
+[options]
+package_dir=
+    =src
+packages=find:
+install_requires =
+    pylint >= 2.7.2
+    pytest >= 6.2.2
+    pytest-flask >= 1.3.0
+    pytest-cov >= 2.11.1
+    wheel >= 0.36.2
+    Flask >= 1.1.2
+    gunicorn >= 20.1.0
+    Flask-Cors >= 3.0.10
+    pymongo >= 4.6.1
+    
+[options.packages.find]
+where=src
